@@ -1,28 +1,31 @@
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'kotlin-kapt'
+}
+
 def version = "1"
 def milestone = "0"
-def build = "0"
+def build = "1"
 
 def versionNum = version.toInteger() * 10000 + milestone.toInteger() * 100 + build.toInteger()
 def versionText = version + "." + milestone + "." + build
 
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
-
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
     defaultConfig {
         applicationId "info.megahard.epshowcase"
         minSdkVersion 15
-        targetSdkVersion 28
+        targetSdkVersion 31
         versionCode versionNum
         versionName versionText
         resValue "string", "app_version", 'v ' + versionText
+        multiDexEnabled true
 
         javaCompileOptions {
             annotationProcessorOptions {
-                arguments = [toothpick_registry_package_name: 'info.megahard.epshowcase.toothpick']
+                arguments = [toothpick_registry_package_name:
+                                     'info.megahard.epshowcase.toothpick']
             }
         }
     }
